Applied node migration 20140326143834
After adding the "language" dimension in the last step, the site was
effectively broken. The existing content was never found, because it
matched none of the possible (configured) dimensions and their values.

So a node migration has been applied to assign all configured the
dimensions and their default values to all nodes not yet having the
default values:

  ./flow node:migrate 20140326143834

Afterwards the site content was again exported via

  ./flow site:export --tidy --package-key Content.Dimensions