Add second dimension, some content
In this change, the dimension configuration is amended with a second
dimension named "system". As with the first dimension that has been
added, a migration needs to be run:

  ./flow node:migrate 20150716212459

This migration adds missing dimensions to existing content.[1] After
is was run, some more content was added and again exported to the
site package.

[1] At the time of writing this, the migration is still in review,
see https://review.typo3.org/41537